fix: drop onnxruntime from client binaries
