+++ Cargo.toml
@@ -1,11 +1,12 @@
 [workspace]
 members = [
     "gaze",
+    "gaze-cli",
     "gaze-core",
     "pam-gaze-core",
     "pam-gaze",
     "pam-gaze-grosshack",
     "gaze-gui",
 ]
-default-members = ["gaze", "pam-gaze", "pam-gaze-grosshack", "gaze-gui"]
+default-members = ["gaze", "gaze-cli", "pam-gaze", "pam-gaze-grosshack", "gaze-gui"]
 resolver = "3"
