+++ Justfile
@@ -25,9 +25,15 @@ default:
 # ── build ─────────────────────────────────────────────────────────────────────
 
 # Build all Rust workspace binaries (release)
+#
+# Split into two invocations so the daemon's `detection` feature on
+# gaze-core does not unify into the client binaries. ONNX Runtime's
+# constructors require AVX2 and crash on older CPUs (issue #14), so the
+# CLI, GUI, and PAM modules must build without it.
 [group("build")]
 build-rust:
-    cargo build --workspace --release
+    cargo build -p gaze --release
+    cargo build -p gaze-cli -p gaze-gui -p pam-gaze -p pam-gaze-grosshack --release
 
 # Compile the SELinux policy module
 [group("build")]
