+++ gaze-cli/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "gaze-cli"
+version = "0.1.0"
+edition = "2024"
+description = "Facial authentication CLI client for Gaze"
+license = "MIT"
+homepage = "https://gaze.gundulabs.com"
+repository = "https://github.com/Gundu-Labs/gaze"
+authors = ["Gundu Labs <maintainers@gundulabs.com>"]
+
+[[bin]]
+name = "gaze"
+path = "src/main.rs"
+
+[dependencies]
+anyhow = "1.0.102"
+clap = { version = "4.6.1", features = ["derive"] }
+console = "0.16.3"
+crossterm = "0.29"
+dialoguer = "0.12"
+futures = "0.3.32"
+gaze-core = { path = "../gaze-core", default-features = false }
+ratatui = "0.30"
+tokio = { version = "1.52", features = ["full", "rt-multi-thread"] }
+zbus = "5.15"
