+++ gaze-core/Cargo.toml
@@ -3,6 +3,10 @@ name = "gaze-core"
 version = "0.1.0"
 edition = "2024"
 
+[features]
+default = ["detection"]
+detection = ["dep:ort", "dep:rusty_scrfd"]
+
 [dependencies]
 anyhow = "1.0.102"
 image = "0.25.10"
@@ -10,8 +14,8 @@ libc = "0.2.186"
 nalgebra = "0.35.0"
 ndarray = "0.17.2"
 opencv = { version = "0.98.2", default-features = false, features = ["imgproc", "videoio"] }
-ort = "2.0.0-rc.12"
-rusty_scrfd = "1.3.1"
+ort = { version = "2.0.0-rc.12", optional = true }
+rusty_scrfd = { version = "1.3.1", optional = true }
 zbus = "5.16.0"
 zvariant = "5.12.0"
 serde = { version = "1.0.228", features = ["derive"] }
