+++ gaze-core/src/capture.rs
@@ -5,7 +5,7 @@ use std::thread;
 use std::time::Duration;
 
 use crate::dbus::CaptureStatus;
-pub use crate::face::{CaptureResult, frame_to_bytes};
+pub use crate::face::CaptureResult;
 
 pub fn init_camera_and_checker(
     device: &str,
