+++ gaze-core/src/lib.rs
@@ -1,7 +1,11 @@
 pub mod camera;
-pub mod capture;
 pub mod config;
 pub mod dbus;
+
+#[cfg(feature = "detection")]
+pub mod capture;
+#[cfg(feature = "detection")]
 pub mod detect;
+#[cfg(feature = "detection")]
 pub mod face;
 pub mod ir;
