+++ gaze/Cargo.toml
@@ -10,7 +10,6 @@ authors = ["Gundu Labs <maintainers@gundulabs.com>"]
 
 [dependencies]
 anyhow = "1.0.102"
-clap = { version = "4.6.1", features = ["derive"] }
 gaze-core = { path = "../gaze-core" }
 image = "0.25.10"
 libc = "0.2.186"
@@ -25,20 +24,11 @@ ureq = "3.3"
 uuid = { version = "1.23.3", features = ["v4"] }
 zbus = "5.16.0"
 zip = "8.6"
-console = "0.16.3"
 tracing = "0.1.44"
 tracing-subscriber = { version = "0.3.23", features = ["env-filter"] }
-dialoguer = "0.12"
-toml = "1.1.2"
 futures = "0.3.32"
 zbus_polkit = "5.0.0"
-crossterm = "0.29"
-ratatui = "0.30.1"
 
 [[bin]]
 name = "gazed"
 path = "src/main.rs"
-
-[[bin]]
-name = "gaze"
-path = "src/bin/cli.rs"
