+++ pam-gaze-core/Cargo.toml
@@ -6,9 +6,8 @@ edition = "2024"
 [lib]
 
 [dependencies]
-gaze-core = { path = "../gaze-core" }
+gaze-core = { path = "../gaze-core", default-features = false }
 libc = "0.2.186"
-opencv = { version = "0.98.2", default-features = false, features = ["imgproc", "videoio"] }
 zbus = "5.16.0"
 tokio = { version = "1.52.3", features = ["rt-multi-thread"] }
 parking_lot = "0.12.5"
