+++ pam-gaze-grosshack/Cargo.toml
@@ -8,7 +8,7 @@ crate-type = ["cdylib"]
 
 [dependencies]
 pam-gaze-core = { path = "../pam-gaze-core" }
-gaze-core = { path = "../gaze-core" }
+gaze-core = { path = "../gaze-core", default-features = false }
 libc = "0.2.186"
 zbus = "5.16.0"
 parking_lot = "0.12.5"
